Updated the code to start containers only for remote

+++ src/test/java/com/fwork/tests/BaseTest.java
@@ -1,6 +1,8 @@
 package com.fwork.tests;
 
 import com.fwork.driver.Driver;
+import com.fwork.enums.ConfigProperties;
+import com.fwork.utils.ReadPropertyFileUtils;
 import com.google.common.util.concurrent.Uninterruptibles;
 import org.testng.ITestContext;
 import org.testng.annotations.*;
@@ -15,16 +17,24 @@ public class BaseTest {
     protected BaseTest() {
 
     }
-
+    public static String runmode = null;
     @BeforeTest
     protected void beforeTest(ITestContext context) {
         context.setAttribute("classname", this.getClass().getSimpleName()); //This is to use in data provider
 
-        //This is to start docker hub and browser containers
         try {
-            Runtime.getRuntime().exec("cmd /c start cmd.exe /K \"cd C:\\Users\\akrpr\\Docker Workspace && docker-compose up -d\"");
-        } catch (IOException e) {
+            runmode = ReadPropertyFileUtils.getValue(ConfigProperties.RUNMODE);
+        } catch (Exception e) {
             throw new RuntimeException(e);
+        }
+
+        //This is to start docker hub and browser containers
+        if(runmode.equalsIgnoreCase("remote")) {
+            try {
+                Runtime.getRuntime().exec("cmd /c start cmd.exe /K \"cd C:\\Users\\akrpr\\Docker Workspace && docker-compose up -d\"");
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
         }
 
         Uninterruptibles.sleepUninterruptibly(15, TimeUnit.SECONDS);
@@ -52,12 +62,16 @@ public class BaseTest {
     }
     @AfterTest
     protected void afterClass() {
-        try {
-            Runtime.getRuntime().exec("cmd /c start cmd.exe /K \"cd C:\\Users\\akrpr\\Docker Workspace && docker-compose down\"");
-            Runtime.getRuntime().exec("cmd /c start cmd.exe /K \"cd C:\\Users\\akrpr\\IdeaProjects\\SeleniumFramework && allure serve\"");
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+
+        if(runmode.equalsIgnoreCase("remote")) {
+            try {
+                Runtime.getRuntime().exec("cmd /c start cmd.exe /K \"cd C:\\Users\\akrpr\\Docker Workspace && docker-compose down\"");
+                //Runtime.getRuntime().exec("cmd /c cmd.exe /K \"cd C:\\Users\\akrpr\\IdeaProjects\\SeleniumFramework && allure serve\"");
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
         }
+
     }
 
 }
